# UniversityStats/UniversityStats.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UniversityStats.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Faculty": {
      "get": {
        "tags": [
          "Faculty"
        ],
        "summary": "Return list of faculties",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacultyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacultyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacultyDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Faculty"
        ],
        "summary": "Post faculty's information to database",
        "requestBody": {
          "description": "Faculty's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Faculty"
        ],
        "summary": "Correct faculty's information by faculty's id",
        "requestBody": {
          "description": "Faculty's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Faculty/{facultyId}": {
      "get": {
        "tags": [
          "Faculty"
        ],
        "summary": "Find faculty's information by faculty's id",
        "parameters": [
          {
            "name": "facultyId",
            "in": "path",
            "description": "Faculty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Faculty"
        ],
        "summary": "Delete faculty by faculty's id",
        "parameters": [
          {
            "name": "facultyId",
            "in": "path",
            "description": "Faculty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Specialty": {
      "get": {
        "tags": [
          "Specialty"
        ],
        "summary": "Return list of specialties",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecialtyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecialtyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecialtyDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Specialty"
        ],
        "summary": "Post specialty's information to database",
        "requestBody": {
          "description": "Specialty's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Specialty"
        ],
        "summary": "Correct specialty's information by specialty's id",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Specialty/{specialtyId}": {
      "get": {
        "tags": [
          "Specialty"
        ],
        "summary": "Return specialty's information by specialty's id",
        "parameters": [
          {
            "name": "specialtyId",
            "in": "path",
            "description": "Specialty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Specialty"
        ],
        "summary": "Delete specialty by specialty's id",
        "parameters": [
          {
            "name": "specialtyId",
            "in": "path",
            "description": "Specialty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FacultyDto": {
        "type": "object",
        "properties": {
          "facultyId": {
            "type": "string",
            "description": "Faculty's id",
            "nullable": true
          },
          "nameFaculty": {
            "type": "string",
            "description": "Faculty's name",
            "nullable": true
          },
          "registrationNumber": {
            "type": "string",
            "description": "University's registration number",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Class for data transfer about faculty's information"
      },
...
      "SpecialtyDto": {
        "type": "object",
        "properties": {
          "specialtyId": {
            "type": "string",
            "description": "Specialty's id",
            "nullable": true
          },
          "nameSpecialty": {
            "type": "string",
            "description": "Specialty's name",
            "nullable": true
          },
          "numberOfGroups": {
            "type": "integer",
            "description": "Total groups in specialty",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Class for data transfer about specialty's information"
      },
...
    }
  }
}